Fix line indentation inconsistencies
I found out that the line indentation in several classes shows differently in GitHub compared to my Eclipse IDE. After looking carefully, there are line indents throughout the codes that are not tab-spaced (if that word makes sense) but instead in multiple single space characters that's why it showing differently. To fix these, I replaced those single space characters with tab-spaced ones to make the indentations consistent.

The most likely reason why this happen is because there are a lot of codes here that I copied and modified from multiple sources, meaning those codes have different approaches in doing a line indentation, thus resulting in this problem. So yeah, it might not be a big deal for others, but for me consistencies is still important even if it's just a minor thing.  <3